Crate srml_consensus

Consensus Module

Overview

The consensus module manages the authority set for the native code. It provides support for reporting offline behavior among validators and logging changes in the validator authority set.

Interface

Dispatchable Functions

report_misbehavior - Report some misbehavior. The origin of this call must be signed.

note_offline - Note that the previous block's validator missed its opportunity to propose a block. The origin of this call must be an inherent.

remark - Make some on-chain remark. The origin of this call must be signed.

set_heap_pages - Set the number of pages in the WebAssembly environment's heap.

set_code - Set the new code.

set_storage - Set some items of storage.

Please refer to the Call enum and its associated variants for documentation on each function.

Public Functions

See the module for details on publicly available functions.

Usage

Prerequisites

To use functionality from the consensus module, implement the specific Trait or function that you are invoking from the module:

impl<T> for consensus::SomeTrait for Module<T> {
	/// required functions and types for trait included here
	/// more comprehensive example included below
}

Alternatively, to set the authorities:

consensus::set_authorities(&[<authorities>]) // example included below

Simple Code Snippet

Set authorities:

<consensus::Module<T>>::set_authorities(&[UintAuthorityId(4), UintAuthorityId(5), UintAuthorityId(6)])

Log changes in the authorities set:

<consensus::Module<T>>::on_finalize(5); // finalize UintAuthorityId(5)

Example from SRML

In the staking module, the consensus::OnOfflineReport is implemented to monitor offline reporting among validators:

impl<T: Trait> consensus::OnOfflineReport<Vec<u32>> for Module<T> {
	fn handle_report(reported_indices: Vec<u32>) {
		for validator_index in reported_indices {
			let v = <session::Module<T>>::validators()[validator_index as usize].clone();
			Self::on_offline_validator(v, 1);
		}
	}
}

In the GRANDPA module, we use srml-consensus to get the set of next_authorities before changing this set according to the consensus algorithm (which does not rotate sessions in the normal way):

let next_authorities = <consensus::Module<T>>::authorities()
			.into_iter()
			.map(|key| (key, 1)) // evenly-weighted.
			.collect::<Vec<(<T as Trait>::SessionKey, u64)>>();

Related Modules

staking: This module uses srml-consensus to monitor offline reporting among validators.

aura: This module does not relate directly to srml-consensus, but serves to manage offline reporting for the Aura consensus algorithm with its own handle_report method.

grandpa: Although GRANDPA does its own voter-set management, it has a mode where it can track consensus, if desired.

References

If you're interested in hacking on this module, it is useful to understand the interaction with substrate/core/inherents/src/lib.rs and, specifically, the required implementation of ProvideInherent to create and check inherents.

Structs

GenesisConfig

InstantFinalityReportVec
A variant of the OfflineReport that is useful for instant-finality blocks.

Module

Enums

Call

RawLog
Logs in this module.

Constants

INHERENT_IDENTIFIER
The identifier for consensus inherents.

Traits

InherentOfflineReport
Describes the offline-reporting extrinsic.

OnOfflineReport
Handling offline validator reports in a generic way.

Trait

Type Definitions

InherentError
The error type used by this inherent.

Key

KeyValue

Log